+++ pydocusaurus/resources/src/envs/variables.jinja2.tsx
@@ -1,0 +1,111 @@
+/**
+ * Environmental variables of this side.
+ * Yuchen Jin, mailto:cainmagi@gmail.com
+ */
+
+import React from "react";
+import Link from "@docusaurus/Link";
+import {useDocsVersion} from "@docusaurus/plugin-content-docs/client";
+
+import InlineIcon from "../components/InlineIcon";
+import mdiDot from "@iconify-icons/mdi/dot";
+
+const docsPluginId = undefined; // Default docs plugin instance
+
+const variables = {
+  repoURL: "https://github.com/{{ user | default("username") | safe }}/{{ package_name | default("pkgname") | safe }}",
+  rawURL: "https://raw.githubusercontent.com/{{ user | default("username") | safe }}/{{ package_name | default("pkgname") | safe }}",
+  sourceVersion: {
+    main: "main",
+  },
+  sourceURIs: {{ source_uris | default({}) | ts_object(indent=2, base_indent=2) }},
+};
+
+{% raw -%}
+const useCurrentSourceVersion = (): string => {
+  const versionHook = useDocsVersion();
+  const versionLabel = versionHook?.label;
+  return (
+    variables.sourceVersion[versionLabel] || variables.sourceVersion["main"]
+  );
+};
+
+export const rawURL = (url: string): string => {
+  return variables.rawURL + "/" + url;
+};
+
+export const repoURL = (url: string | undefined = undefined): string => {
+  return url ? variables.repoURL + "/" + url : variables.repoURL;
+};
+
+export const releaseURL = (ver: string | undefined = undefined): string => {
+  const _ver = ver?.toLowerCase() === "next" ? "main" : ver;
+  const version = variables.sourceVersion[_ver] || useCurrentSourceVersion();
+  if (version === "main" || _ver === "main") {
+    return variables.repoURL + "/releases/latest";
+  }
+  return variables.repoURL + "/releases/tag/" + version;
+};
+
+export const rootURL = (url: string): string => {
+  const currentSourceVersion = useCurrentSourceVersion();
+  return variables.repoURL + "/blob/" + currentSourceVersion + "/" + url;
+};
+
+const getURIByVersionPath = (path: string, ver: string): string => {
+  const routes = typeof path === "string" ? path.trim() : "";
+  if (routes.length === 0) {
+    return path;
+  }
+  const currentURI = variables.sourceURIs[ver] || variables.sourceURIs["main"];
+  return currentURI[path] || path;
+};
+{%- endraw %}
+
+export const sourceURL = (url: string): string => {
+  const currentSourceVersion = useCurrentSourceVersion();
+  return (
+    variables.repoURL +
+    "/blob/" +
+    currentSourceVersion +
+    "/{{ package_name | default("pkgname") | safe }}/" +
+    getURIByVersionPath(url, currentSourceVersion)
+  );
+};
+
+{% raw -%}
+export const demoURL = (url?: string): string => {
+  const currentSourceVersion = useCurrentSourceVersion();
+  if (!url) {
+    return variables.repoURL + "/blob/" + currentSourceVersion + "/usage.py";
+  }
+  return (
+    variables.repoURL + "/blob/" + currentSourceVersion + "/examples/" + url
+  );
+};
+
+export type SourceLinkProps = {
+  url: string;
+  children: React.ReactNode;
+};
+
+export const SourceLink = ({url, children}: SourceLinkProps): JSX.Element => {
+  return (
+    <Link to={sourceURL(url)} className="noline">
+      {children}
+    </Link>
+  );
+};
+
+export type SplitterProps = {
+  padx?: string;
+};
+
+export const Splitter = ({padx = "0"}: SplitterProps): JSX.Element => {
+  return (
+    <span style={{padding: "0 " + padx}}>
+      <InlineIcon icon={mdiDot} />
+    </span>
+  );
+};
+{%- endraw %}